Consolidate logging setup
